+++ packages/EasyApiPlatform/tests/Filters/AdvancedSearchFilterTest.php
@@ -9,9 +9,9 @@ use ApiPlatform\Metadata\Exception\InvalidArgumentException;
 use ApiPlatform\Metadata\Get;
 use Doctrine\Persistence\ManagerRegistry;
 use EonX\EasyApiPlatform\Filters\AdvancedSearchFilter;
-use EonX\EasyApiPlatform\Tests\Fixtures\App\Case\AdvancedSearchFilter\ApiResource\Dummy;
-use EonX\EasyApiPlatform\Tests\Fixtures\App\Case\AdvancedSearchFilter\ApiResource\RelatedDummy;
-use EonX\EasyApiPlatform\Tests\Fixtures\App\Case\AdvancedSearchFilter\Serializer\NameConverter\CustomConverter;
+use EonX\EasyApiPlatform\Tests\Fixtures\App\AdvancedSearchFilter\ApiResource\Dummy;
+use EonX\EasyApiPlatform\Tests\Fixtures\App\AdvancedSearchFilter\ApiResource\RelatedDummy;
+use EonX\EasyApiPlatform\Tests\Fixtures\App\AdvancedSearchFilter\Serializer\NameConverter\CustomConverter;
 use Prophecy\Argument;
 use Prophecy\Prophet;
