+++ packages/EasyApiPlatform/tests/Fixtures/app/config/packages/doctrine.php
@@ -20,20 +20,20 @@ return static function (DoctrineConfig $doctrineConfig, ContainerConfigurator $c
         ->entityManager('default');
 
     $entityManager->mapping('AppCaseAdvancedSearchFilterApiResource')
-        ->dir(param('kernel.project_dir') . '/src/Case/AdvancedSearchFilter/ApiResource')
+        ->dir(param('kernel.project_dir') . '/src/AdvancedSearchFilter/ApiResource')
         ->isBundle(false)
-        ->prefix('EonX\EasyApiPlatform\Tests\Fixtures\App\Case\AdvancedSearchFilter\ApiResource')
+        ->prefix('EonX\EasyApiPlatform\Tests\Fixtures\App\AdvancedSearchFilter\ApiResource')
         ->type('attribute');
 
     $entityManager->mapping('AppCaseCustomPaginatorApiResource')
-        ->dir(param('kernel.project_dir') . '/src/Case/CustomPaginator/ApiResource')
+        ->dir(param('kernel.project_dir') . '/src/CustomPaginator/ApiResource')
         ->isBundle(false)
-        ->prefix('EonX\EasyApiPlatform\Tests\Fixtures\App\Case\CustomPaginator\ApiResource')
+        ->prefix('EonX\EasyApiPlatform\Tests\Fixtures\App\CustomPaginator\ApiResource')
         ->type('attribute');
 
     $entityManager->mapping('AppCaseReadListenerApiResource')
-        ->dir(param('kernel.project_dir') . '/src/Case/ReadListener/ApiResource')
+        ->dir(param('kernel.project_dir') . '/src/ReadListener/ApiResource')
         ->isBundle(false)
-        ->prefix('EonX\EasyApiPlatform\Tests\Fixtures\App\Case\ReadListener\ApiResource')
+        ->prefix('EonX\EasyApiPlatform\Tests\Fixtures\App\ReadListener\ApiResource')
         ->type('attribute');
 };
