+++ packages/EasyApiPlatform/tests/Fixtures/app/src/AdvancedSearchFilter/ApiResource/Dummy.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace EonX\EasyApiPlatform\Tests\Fixtures\App\Case\AdvancedSearchFilter\ApiResource;
+namespace EonX\EasyApiPlatform\Tests\Fixtures\App\AdvancedSearchFilter\ApiResource;
 
 use ApiPlatform\Doctrine\Orm\Filter\SearchFilter;
 use ApiPlatform\Metadata\ApiFilter;
@@ -68,7 +68,7 @@ class Dummy
     private ?string $nameConverted = null;
 
     /**
-     * @var \Doctrine\Common\Collections\Collection<int, \EonX\EasyApiPlatform\Tests\Fixtures\App\Case\AdvancedSearchFilter\ApiResource\RelatedDummy> Several dummies
+     * @var \Doctrine\Common\Collections\Collection<int, \EonX\EasyApiPlatform\Tests\Fixtures\App\AdvancedSearchFilter\ApiResource\RelatedDummy> Several dummies
      */
     #[ORM\ManyToMany(targetEntity: RelatedDummy::class)]
     private Collection $relatedDummies;
@@ -169,7 +169,7 @@ class Dummy
     }
 
     /**
-     * @return \Doctrine\Common\Collections\Collection<int, \EonX\EasyApiPlatform\Tests\Fixtures\App\Case\AdvancedSearchFilter\ApiResource\RelatedDummy>
+     * @return \Doctrine\Common\Collections\Collection<int, \EonX\EasyApiPlatform\Tests\Fixtures\App\AdvancedSearchFilter\ApiResource\RelatedDummy>
      */
     public function getRelatedDummies(): Collection
     {
