+++ packages/EasyApiPlatform/tests/Fixtures/app/src/AdvancedSearchFilter/ApiResource/RelatedOwnedDummy.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace EonX\EasyApiPlatform\Tests\Fixtures\App\Case\AdvancedSearchFilter\ApiResource;
+namespace EonX\EasyApiPlatform\Tests\Fixtures\App\AdvancedSearchFilter\ApiResource;
 
 use ApiPlatform\Metadata\ApiResource;
 use Doctrine\DBAL\Types\Types;
@@ -9,7 +9,7 @@ use Doctrine\ORM\Mapping as ORM;
 
 #[ApiResource(types: ['https://schema.org/Product'])]
 #[ORM\Entity]
-class RelatedOwningDummy
+class RelatedOwnedDummy
 {
     #[ORM\Column(type: Types::INTEGER)]
     #[ORM\GeneratedValue(strategy: 'AUTO')]
@@ -19,8 +19,9 @@ class RelatedOwningDummy
     #[ORM\Column(type: Types::STRING, nullable: true)]
     private ?string $name = null;
 
-    #[ORM\OneToOne(mappedBy: 'relatedOwningDummy', targetEntity: Dummy::class, cascade: ['persist'])]
-    private ?Dummy $ownedDummy = null;
+    #[ORM\JoinColumn(nullable: false)]
+    #[ORM\OneToOne(inversedBy: 'relatedOwnedDummy', targetEntity: Dummy::class, cascade: ['persist'])]
+    private Dummy $owningDummy;
 
     public function getId(): int
     {
@@ -32,9 +33,12 @@ class RelatedOwningDummy
         return $this->name;
     }
 
-    public function getOwnedDummy(): ?Dummy
+    /**
+     * Get owning dummy.
+     */
+    public function getOwningDummy(): ?Dummy
     {
-        return $this->ownedDummy;
+        return $this->owningDummy;
     }
 
     public function setId(int $id): void
@@ -47,12 +51,8 @@ class RelatedOwningDummy
         $this->name = $name;
     }
 
-    public function setOwnedDummy(?Dummy $ownedDummy): void
+    public function setOwningDummy(Dummy $owningDummy): void
     {
-        $this->ownedDummy = $ownedDummy;
-
-        if ($this->ownedDummy !== null && $this !== $this->ownedDummy->getRelatedOwningDummy()) {
-            $this->ownedDummy->setRelatedOwningDummy($this);
-        }
+        $this->owningDummy = $owningDummy;
     }
 }
